updated files

+++ js/sign-up.js
@@ -5,12 +5,16 @@ warningParagraph.textContent = 'Email already exists';
 warningParagraph.style.color = 'rgb(219, 43, 43)';
 warningParagraph.style.fontSize = '17px';
 
+const secondPassword = document.querySelector('.confirmed-password');
+const unmatchPassword = document.createElement('p');
+unmatchPassword.textContent = 'Password does not match';
+unmatchPassword.style.color = 'rgb(219, 43, 43)';
+
+
 const getSavedUser = () => {
     const userJSON = localStorage.getItem('user');
     return userJSON ? JSON.parse(userJSON) : [];
 }
-
-
 
 let user = getSavedUser();
 
@@ -20,20 +24,28 @@ userSignUp.addEventListener('submit', (e) => {
     const lastName = e.target.elements.lastName.value;
     const email = e.target.elements.email.value;
     const password = e.target.elements.password.value;
+    const confirmedPassword = e.target.elements.confirmedPassword.value;
 
-    if (user.find((object) => object.email === email)) {
-        emailTitle.appendChild(warningParagraph);
+    if (password === confirmedPassword) {
+        if (user.find((object) => object.email === email)) {
+            emailTitle.appendChild(warningParagraph);
+        }
+        else {
+            console.log(user.includes(email));
+            user.push({
+                firstName: firstName,
+                lastName: lastName,
+                email: email,
+                password: password
+            });
+            localStorage.setItem('user', JSON.stringify(user));
+            location.assign('/html/index.html');
+        }
     }
     else {
-        console.log(user.includes(email));
-        user.push({
-            firstName: firstName,
-            lastName: lastName,
-            email: email,
-            password: password
-        });
-        localStorage.setItem('user', JSON.stringify(user));
-        location.assign('/html/index.html');
+        secondPassword.appendChild(unmatchPassword);
     }
-
 });
+
+
+
